GraphNodes are actually node/children so we can write them out very much like Flipped nodes.
Purge unused segment code.
