+++ osm-jigsaw-parser/src/main/scala/steps/BuildGraph.scala
@@ -1,23 +1,23 @@
 package steps
 
 import areas.AreaComparison
-import graphing.{GraphBuilder, GraphWriter}
+import graphing.GraphBuilder
 import input.AreaReading
 import model.{Area, GraphNode}
 import org.apache.logging.log4j.scala.Logging
-import output.OutputFiles
+import output.{GraphWriting, OutputFiles}
 import progress.ProgressCounter
 
 import java.io.{BufferedOutputStream, FileOutputStream}
 
-class BuildGraph extends OutputFiles with AreaReading with Segmenting with AreaComparison with Logging {
+class BuildGraph extends OutputFiles with AreaReading with AreaComparison with GraphWriting with Logging {
 
   def buildGraph(extractName: String) = {
     val areas = readAreasFromPbfFile(areasFilePath(extractName))
 
     logger.info("Building graph")
 
-    val headArea = Area(-1L, null, (0,0,0,0), area = 0)
+    val headArea = Area(-1L, null, (0, 0, 0, 0), area = 0)
     val graph = new GraphBuilder().buildGraph(headArea, areas)
     writeGraph(graph, new FileOutputStream(graphFile(extractName)))
     logger.info("Done")
@@ -26,8 +26,7 @@ class BuildGraph extends OutputFiles with AreaReading with Segmenting with AreaC
   private def writeGraph(head: GraphNode, outputStream: FileOutputStream) = {
     logger.info("Writing graph to disk")
     val output = new BufferedOutputStream(outputStream)
-    val counter = new ProgressCounter(10000)
-    new GraphWriter().export(head, output, None, counter)
+    outputGraph(head, output)
     output.flush()
     output.close()
   }
